chromTFR: Quantification of Chromatin Accessibility Signatures in TFBS

chromTFR quantifies transcription factor footprints in chromatin accessibility data using the annotation of the methylTFR package. Tn5 insertion densities replace methylation levels, and the expected profile of a motif is derived either from the GC bin composition of its binding sites or from a k-mer model of the Tn5 sequence preference. The package computes bias corrected deviation scores per sample and motif, draws the underlying footprints and summarises the deviations as a heatmap.

See also

Author

Maintainer: Irem B. Gunduz irembgunduz@gmail.com (ORCID)

Authors:

Fabian Muller fabian.mueller@uni-saarland.de (ORCID)